Comment out what's not implemented in turnkey
https://github.com/douglasnaphas/madliberation/issues/272

The "turnkey env," "turnkey setup," or "turnkey" refers to
https://github.com/douglasnaphas/madliberation, a repo attempting to
deploy passover.lol using the AWS CDK and my "turnkey" principles of
deploying a usable site varying only the AWS creds in GitHub Secrets.

This commit strips the integration test script down to only test what is
implemented in the turnkey env, where this itest project is a submodule.