Log in and change password, then join
https://github.com/douglasnaphas/madliberation/issues/274

This sets up the integration test to create a user (in setup, via the
SDK), then log in as the user. This validates that the user pool is
usable.

The last implemented piece is trying to join the seder as Player 2. This
should fail, because /get-cookies is not set up under turnkey yet.